+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tracking skipped items:
   - Information message shows amount of created and skipped items.
 
+### Removed
+- Information message for each separate skipped file.
+
 
 ## [2.1.2] – 2025-10-02
 ### Added