+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Tracking skipped items:
   - Information message shows amount of created and skipped items.
-  - Button to view what items got skipped.
+  - Button to view what items got skipped, if any got skipped.
 
 ### Removed
 - Information message for each separate skipped file.